Actualizacion Funcionamiento a través de curl correcto y se muestran datos en pestañas del navegador. /***************************************************************************/

+++ MonitorEventos/multiproyecto/monitorApi/grails-app/domain/fiuba/materia7510/aplicacion/proveedor/Ticket.groovy
@@ -1,12 +1,12 @@
 package fiuba.materia7510.aplicacion.proveedor
 
 
-class Ticket{
+class Ticket {
 
 	
 	Estado 	estado
 	
-	Integer 	codigo
+	Long 	codigo
 	
 	String 		titulo
 	String 		propietario
